chore: refactored structure and names
