+++ meowtg/plugin/versionInfo.ts
